% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{GetSecondaryDirectoryFile}
\alias{GetSecondaryDirectoryFile}
\title{GetSecondaryDirectoryFile}
\usage{
GetSecondaryDirectoryFile(xx)
}
\arguments{
\item{xx}{随意填入数字}
}
\value{
filename
}
\description{
验证二级文件夹里面是否有文件，无返回NA值，有为TRUE。最后返回二级文件夹里面的第一个文件绝对路径
}
