% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{GetSecondaryDirectoryFile}
\alias{GetSecondaryDirectoryFile}
\title{Title 验证二级文件夹里面是否有文件，无返回NA值，为TRUE。最后返回二级文件夹文件}
\usage{
GetSecondaryDirectoryFile(xx)
}
\arguments{
\item{xx}{}
}
\value{

}
\description{
Title 验证二级文件夹里面是否有文件，无返回NA值，为TRUE。最后返回二级文件夹文件
}
